moving jumping gains into separate class to reduce clutter
